+++ src/com/company/Operation.java
@@ -1,5 +1,5 @@
 package com.company;
 
 public interface Operation {
-    double calculate(double a, double b);
+    double calculate(int a, int b);
 }
